+++ CHANGELOG.md
@@ -16,6 +16,9 @@ All changes and updates, that are relevant for a user will be documented here
 
 ### Features
 
+- Automatically check for new MolarTracker releases on startup and every
+  24 h; if a newer version is available, a dialog is shown with a link to
+  the releases page and a "don't show again for this version" option
 - Add possibility to add fees when creating new stock or cash (deposit/withdrawal) transactions
 - Add possibility to persist different log levels for different categories
 - Show full exception stack trace when an exception occurs